Total time needs to be double, fix time calculation

+++ src/game_time.cpp
@@ -2,7 +2,7 @@
 
 namespace arena
 {
-	GameTime::GameTime(const float32 delta, const float32 total) : m_delta(delta),
+	GameTime::GameTime(const float32 delta, const float64 total) : m_delta(delta),
 																   m_total(total)
 	{
 	}
